store fontsize as a double in memory

+++ lib/epub/parser/epub_parser_worker.dart
@@ -237,7 +237,7 @@ class EpubParserWorker {
             cssParser.parseCss(message['css']);
             break;
           case _fontSize:
-            ElementStyle.defaultFontSize = message['fontSize'];
+            ElementStyle.defaultFontSize = message['fontSize'].toDouble();
             break;
           case _openBook:
             parser.openBook(message['book']);
