+++ lib/epub/styles/element_style.dart
@@ -10,7 +10,7 @@ class ElementStyle extends Style {
   late TextStyle textStyle;
 
   bool? isDropCaps;
-  static int defaultFontSize = 12;
+  static double defaultFontSize = 12;
 
   // Character: (superscript subscript)
   static const Map<String, ({String sup, String sub})> unicodeMap = {
@@ -59,7 +59,7 @@ class ElementStyle extends Style {
 
   void getTextStyle(XmlNode element) {
     final String? fontFamily = _parser.getFontAttribute(element, this, "font-family")?.replaceAll('"', '');
-    final double? fontSize   = defaultFontSize.toDouble();
+    final double? fontSize   = defaultFontSize;
     final String? fontStyle  = _parser.getStringAttribute(element, this, "font-style");
     final String? fontWeight = _parser.getStringAttribute(element, this, "font-weight");
     final String? fontDecoration  = _parser.getStringAttribute(element, this, "text-decoration");
